+++ src/main/java/io/github/astrapi69/comparator/ComparatorExtensions.java
@@ -18,7 +18,7 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.comparators;
+package io.github.astrapi69.comparator;
 
 /**
  * The class {@link ComparatorExtensions} have helper methods for comparators.