+++ src/test/java/io/github/astrapi69/comparator/ComparatorExtensionsTest.java
@@ -18,12 +18,14 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.comparators;
+package io.github.astrapi69.comparator;
 
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNull;
 import static org.testng.AssertJUnit.assertTrue;
 
+import io.github.astrapi69.comparator.ComparatorExtensions;
+import io.github.astrapi69.comparator.SortOrder;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;