+++ src/test/java/io/github/astrapi69/comparator/SortOrderComparatorTest.java
@@ -18,12 +18,14 @@
  * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.comparators;
+package io.github.astrapi69.comparator;
 
 import static org.testng.AssertJUnit.assertTrue;
 
 import java.util.Comparator;
 
+import io.github.astrapi69.comparator.SortOrder;
+import io.github.astrapi69.comparator.SortOrderComparator;
 import org.testng.annotations.Test;
 
 /**